Move kernel info schemes to sys:
