+++ kernel/schemes/mod.rs
@@ -1,5 +1,3 @@
-/// Context scheme
-pub mod context;
 /// Debug scheme
 pub mod debug;
 /// Disk scheme
@@ -10,15 +8,9 @@ pub mod display;
 pub mod env;
 /// Init Filesystem
 pub mod initfs;
-/// Interrupt scheme
-pub mod interrupt;
-/// Memory scheme
-pub mod memory;
 /// Pipes
 pub mod pipe;
 /// Psuedoterminals
 pub mod pty;
-/// Logging scheme
-pub mod syslog;
-/// Tests
-pub mod test;
+/// Sys scheme
+pub mod sys;
