+++ kernel/schemes/sys/test/get_slice.rs
